New feature: allow-listing of issues
A client wants all the time to be reported in broad categories instead
of mapped to exact Jira tickets.  For this purpose there's a Jira epic
with a number of child issues to be used for time-reporting purposes
only, so I want those to be the only ones that get time reports.

This is achieved by adding an

    include = ...

configuration statement that lists all the tickets used for
time-reporting purposes.